All parameters are read from text file
There were two major changes in this update. Firstly, the file requires
compilation with C++11. Makefile has been altered to support this.
Secondly, all of the parameters are read from the file "/params.txt".
The function Block::readParameters() takes the filename and reads
through the file line by line, cutting each line into substring by
deliminaion of an empty space ' '. If the
first word matches one of the parameters, the second element is taken as
its value. There is no error checking, so be utmost cautious when
writing/modifying the parameters.

In other news, the variables tStop and dt are made members of the Block
class. Logically, this might not have been the best idea, but it was the
easiest solution for reading those parameters from file. numBlocks was
changed from const to non-const, which is a very bad idea. The reason
for why this nevertheless was done, was to make it possible to read
numBlocks from file, as const must be intitialized by the constructor. A
cleanup is needed.