Fixed const by adding parameter class
Last problem introduced the ability to read parameters from file.
Unfortunately, this was not the most elegant solution, as the Blocks
class should ideally only handle computiation, not file I/O. Even worse,
numBlocks was no longer const, as its value has the be set in the
constructor, before any file is read. The parameter class was written to
remedy these issues. It is very simple, acting as a mediator between the
parameter file and the Block class. Currently its only purpose in life
is to read the parameter file and pass them on to the Block class, such
that numBlocks may be const.

As a side note, still working on why Emacs screws up the indentation.