Fix timer bug and make connector struct
With help from Adam, a bug was noticed with the timers. The timers were
set to m_t at initilization, but since the m_time_limit is constant, m_t
will always be greater than m_time_limit after a few time steps. Thus,
the timers are not set to 0.

Scratching the last idea with a connector class, a connector struct is
implemented instead. It is very simple but serves its purpose well by
making the code more organized. However, since m_states are no longer
set, a function Blocks::fillStatesArray was written to fill m_states
with the states of the connectors before they are dumped to file.