Rewrite entire class system
Major changes are introduced in this commit. The class system as been
completely rewritten to make it more manageable. Only the Params class
is left unharmed. The old Blocks class has been split into two - a
System class to manage the variables, simulation and integration, and a
Block class representing each individual block.

The System class contains most of the arrays that Blocks contained,
adding only a new m_blocks vector that contains the blocks. For the most
part, the Block class access these arrays by directly altering their
values through pointers. System only calls the Block class for force
calculation and dumping the connectors/micro-junctions to m_states.

As mentioned, the Block class manipulates the System's arrays by
dereferencing pointers. This method is risky and errorprone, but fast
and greatly simplifies the interface between Block and System. As with
most of C++, if you get wierd behaviour, check the pointers.

The Block class supports a variable number of connectors. This number is
set in params.txt with numConnectors. The connectors are contained in an
array, m_connectors.

Since the first and last block compute the forces differently from the
rest, two derived classes - FirstBlock and LastBlock - are implemented.
Their constructor is the same as that of Block, while calculateForces()
is different. To accomplish this, Block's calculateForces is virtual