Is now 2D. Not working
This commit introduces major changes. First, a new Vector class has been
implemented to ease vector calculations. Secondly, the modell is now 2D.
To make array iteration as quick as possible, the arrays are not 2D, but
1D of the form [num_rows*num_columns], indexed as
[current_column*num_rows + current_row].

Alas, it isn't working as intended. To be fixed. Will add unique_ptr to
stop pesky pointer errors.