Fix several bugs. The modell is now working
After hours of bug hunting, the 2D-modell is now working. The bugs were
some rouge pointers, wrong indecies and wrong sign in operations.

The 2D modell is a grid indexed as a matrix, starting (in memory) from
[0,0] in the to left and ending in [numBlocksY,numBlocksX]. However, to
make the code more efficient, only the blockstructure is implemented in
this manner, the rest are 1D-arrays indexed as [y*numBlocksX+x].

File I/O has been moved to System to make the code more managable. Note
that fillStatesArray still exists, sadly, and must be called before the
states are written to file.

Since the data is now 2D, the python script has been changed
accordingly. To visualise the forces, the plot axis plots the Y-X grid
at a given time step which is controlled by a slider. An animation is
probably more convient, but may be slower.