Fix the optimisation bug
After hours of painstakingly debugging the code, I finally found the
source. I do not understand why the compiler made the optimisation that
it did, but I some ideas.

The error stems from how the BottomBlock accessed m_pFrictionforce. It
was only accessed as *(m_pFrictionforce+i), and by some strange reason,
the compiler assumed that i=/=0. When i was 0, *(m_pFrictionforce+0)
gave a value of some random location in memory, causing the simulation
to overflow. This bug is extremely persisent, as the compiler seems to
ignore volatile and pragma directives.

The bug was solved by adding a never-true if-statement in the
BottomBlock's deconstructor which when true prints the value
of *m_pFrictionforce. This keeps the compiler from messing with the
pointer.